move equation `Eval` impl to `math.rs`
